allow click anywhere on "add" dropdown line
It's obnoxious to have to have your mouse exactly over the exact text of
the option in the "+ Add" dropdown in order to make it work (as opposed
to clicking anywhere on that option's line). This should fix that.
Please excuse my committing this to an unrelated branch. :see_no_evil: