Remove commented less import.
I asked @tmsp1 to comment this so I can check if it needs to be
moved elsewhere, or whether it can simply be removed. Turns out
the bootstrap variables are already included in styles/main.less
and also in every other product where this is re-used, so deleting it
is the right thing to do.